requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9

# NLP
sentence-transformers==2.7.0
numpy>=1.26
pyyaml>=6.0

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.1

# Observability
prometheus-client==0.20.0
sentry-sdk[fastapi]==2.3.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-timeout==2.3.1
httpx==0.27.0
